requirements.txt
streamlit
requests
openai
numpy
sympy
python-dotenv